// src/api/model/school.ts
/**
 * Generated by orval v6.26.0 🍺
 * Do not edit manually.
 * student-hub-api
 * Api Documentation for StudentHub API
 * OpenAPI spec version: 0.0
 */

export interface School {
  rank: number;
  schoolDistrict: string;
  schoolId: number;
  schoolMunicipality: string;
  schoolName: string;
  totalCorrectAnswers: number;
  totalQuestionsAnswered: number;
  totalUsers: number;
}